requirements.txt
openai
supabase
python-dotenv
Pillow
requests